A dynamic web page with MySQL database
A dynamic web page with MySQL database connected through JDBC and JSP.
For the front-end HTML and CSS were used.

A web application to create a secure Customer account and store it in
the MySQL database. Transfer funds from one account to another with a
Service Charge. An offline Passbook and balance verification capability.